# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Recruiter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  open: 'open',
  closed: 'closed',
  draft: 'draft',
} as const;

export interface Job {
  id: number;
  title: string;
  department: string;
  location: string;
  status: JobStatus;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  requirements?: string | null;
  /** @nullable */
  employmentType?: string | null;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  applicantCount?: number;
  createdAt: string;
  updatedAt?: string;
}

export type JobInputStatus = typeof JobInputStatus[keyof typeof JobInputStatus];


export const JobInputStatus = {
  open: 'open',
  closed: 'closed',
  draft: 'draft',
} as const;

export interface JobInput {
  /** @minLength 1 */
  title: string;
  department: string;
  location: string;
  status?: JobInputStatus;
  description?: string;
  requirements?: string;
  employmentType?: string;
  salaryMin?: number;
  salaryMax?: number;
}

export type JobUpdateStatus = typeof JobUpdateStatus[keyof typeof JobUpdateStatus];


export const JobUpdateStatus = {
  open: 'open',
  closed: 'closed',
  draft: 'draft',
} as const;

export interface JobUpdate {
  title?: string;
  department?: string;
  location?: string;
  status?: JobUpdateStatus;
  description?: string;
  requirements?: string;
  employmentType?: string;
  salaryMin?: number;
  salaryMax?: number;
}

export interface Candidate {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  currentTitle?: string | null;
  /** @nullable */
  currentCompany?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  resumeSummary?: string | null;
  tags?: string[];
  createdAt: string;
  updatedAt?: string;
}

export interface CandidateInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string;
  currentTitle?: string;
  currentCompany?: string;
  location?: string;
  linkedinUrl?: string;
  githubUrl?: string;
  resumeSummary?: string;
  tags?: string[];
}

export interface CandidateUpdate {
  name?: string;
  email?: string;
  phone?: string;
  currentTitle?: string;
  currentCompany?: string;
  location?: string;
  linkedinUrl?: string;
  githubUrl?: string;
  resumeSummary?: string;
  tags?: string[];
}

export type ApplicationStage = typeof ApplicationStage[keyof typeof ApplicationStage];


export const ApplicationStage = {
  applied: 'applied',
  screening: 'screening',
  interview: 'interview',
  offer: 'offer',
  hired: 'hired',
  rejected: 'rejected',
} as const;

export interface Application {
  id: number;
  jobId: number;
  candidateId: number;
  stage: ApplicationStage;
  /** @nullable */
  fitScore?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  rejectionReason?: string | null;
  /** @nullable */
  outreachStatus?: string | null;
  createdAt: string;
  updatedAt?: string;
  job?: Job;
  candidate?: Candidate;
}

export type ApplicationInputStage = typeof ApplicationInputStage[keyof typeof ApplicationInputStage];


export const ApplicationInputStage = {
  applied: 'applied',
  screening: 'screening',
  interview: 'interview',
  offer: 'offer',
  hired: 'hired',
  rejected: 'rejected',
} as const;

export interface ApplicationInput {
  jobId: number;
  candidateId: number;
  stage?: ApplicationInputStage;
  fitScore?: number;
  notes?: string;
}

export type ApplicationUpdateStage = typeof ApplicationUpdateStage[keyof typeof ApplicationUpdateStage];


export const ApplicationUpdateStage = {
  applied: 'applied',
  screening: 'screening',
  interview: 'interview',
  offer: 'offer',
  hired: 'hired',
  rejected: 'rejected',
} as const;

export interface ApplicationUpdate {
  stage?: ApplicationUpdateStage;
  fitScore?: number;
  notes?: string;
  rejectionReason?: string;
  outreachStatus?: string;
}

export type PipelineStatsStageBreakdown = {[key: string]: number};

export interface PipelineStats {
  totalJobs: number;
  openJobs: number;
  totalCandidates: number;
  activeApplications: number;
  hiredThisMonth: number;
  /** @nullable */
  avgFitScore: number | null;
  stageBreakdown: PipelineStatsStageBreakdown;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  candidateName?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  createdAt: string;
}

export interface FunnelStage {
  stage: string;
  count: number;
  percentage: number;
}

export type ListJobsParams = {
status?: ListJobsStatus;
};

export type ListJobsStatus = typeof ListJobsStatus[keyof typeof ListJobsStatus];


export const ListJobsStatus = {
  open: 'open',
  closed: 'closed',
  draft: 'draft',
} as const;

export type ListCandidatesParams = {
jobId?: number;
stage?: string;
};

export type ListApplicationsParams = {
jobId?: number;
candidateId?: number;
stage?: string;
};

export type GetRecentActivityParams = {
limit?: number;
};

export type GetPipelineFunnelParams = {
jobId?: number;
};