# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Recruiter API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  candidateName?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  createdAt: Date;
}