# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Recruiter API
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationInputStage } from './applicationInputStage';

export interface ApplicationInput {
  jobId: number;
  candidateId: number;
  stage?: ApplicationInputStage;
  fitScore?: number;
  notes?: string;
}