# lib/api-zod/src/generated/types/applicationUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Recruiter API
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationUpdateStage } from './applicationUpdateStage';

export interface ApplicationUpdate {
  stage?: ApplicationUpdateStage;
  fitScore?: number;
  notes?: string;
  rejectionReason?: string;
  outreachStatus?: string;
}