# lib/api-zod/src/generated/types/candidate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Recruiter API
 * OpenAPI spec version: 0.1.0
 */

export interface Candidate {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  currentTitle?: string | null;
  /** @nullable */
  currentCompany?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  resumeSummary?: string | null;
  tags?: string[];
  createdAt: Date;
  updatedAt?: Date;
}