# lib/api-zod/src/generated/types/candidateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Recruiter API
 * OpenAPI spec version: 0.1.0
 */

export interface CandidateInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string;
  currentTitle?: string;
  currentCompany?: string;
  location?: string;
  linkedinUrl?: string;
  githubUrl?: string;
  resumeSummary?: string;
  tags?: string[];
}