# lib/api-zod/src/generated/types/funnelStage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Recruiter API
 * OpenAPI spec version: 0.1.0
 */

export interface FunnelStage {
  stage: string;
  count: number;
  percentage: number;
}